Điều chỉnh phần xóa ở Categories, Sizes, Colors

+++ ShoesStore.Application/Features/Categories/Commands/DeleteCategory/DeleteCategoryCommandHander.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using ShoesStore.Application.Common.Interfaces;
 using ShoesStore.Domain.Entities.Data;
@@ -19,28 +20,30 @@ namespace ShoesStore.Application.Features.Categories.Commands.DeleteCategory
 
         public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
-            try
+            var categoryDb = await _unitOfWork.Repository<Category>().GetByIdAsync(request.Id);
+            var productsRepo = _unitOfWork.Repository<Product>();
+
+            var isHadproductsRepo = await productsRepo.GetQueryable()
+                                    .AnyAsync(p => p.ProductCategories.Any(pc => pc.CategoryId == request.Id), cancellationToken);
+
+            if (isHadproductsRepo)
             {
-                var categoryDb = await _unitOfWork.Repository<Category>().GetByIdAsync(request.Id);
-
-                if (categoryDb == null)
-                {
-                    throw new KeyNotFoundException($"Category with ID {request.Id} not found.");
-                }
-
-                await _unitOfWork.Repository<Category>().DeleteAsync(categoryDb, cancellationToken);
-
-                await _unitOfWork.SaveChangesAsync(cancellationToken);
-
-                _logger.LogInformation("Xóa danh mục giày thành công với Id: {Id}", request.Id);
-
-                return Unit.Value;
+                // Nếu đang được sử dụng, ném ra một ngoại lệ với thông báo rõ ràng.
+                throw new InvalidOperationException("Không thể xóa danh mục này vì đang có sản phẩm sử dụng.");
             }
-            catch (Exception ex)
+
+            if (categoryDb == null)
             {
-                _logger.LogError(ex, "Lỗi khi xóa danh mục giày với Id: {Id}", request.Id);
-                throw new ApplicationException("Lỗi khi xóa danh mục giày", ex);
+                throw new KeyNotFoundException($"Danh mục với id {request.Id} không tồn tại.");
             }
+
+            await _unitOfWork.Repository<Category>().DeleteAsync(categoryDb, cancellationToken);
+
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Xóa danh mục giày thành công với Id: {Id}", request.Id);
+
+            return Unit.Value;
         }
     }
 }
