+++ ShoesStore.Application/Features/Colors/Commands/DeleteColor/DeleteColorCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using ShoesStore.Application.Common.Interfaces;
 using ShoesStore.Domain.Entities.Data;
@@ -18,28 +19,30 @@ namespace ShoesStore.Application.Features.Colors.Commands.DeleteColor
 
         public async Task<Unit> Handle(DeleteColorCommand request, CancellationToken cancellationToken)
         {
-            try
+            var colorDb = await _unitOfWork.Repository<Color>().GetByIdAsync(request.Id);
+            var prodVarRepo = _unitOfWork.Repository<ProductVariant>();
+
+            var isHadProdVar = await prodVarRepo.GetQueryable()
+                .AnyAsync(pv => pv.ColorId == request.Id, cancellationToken);
+
+            if (isHadProdVar)
             {
-                var colorDb = await _unitOfWork.Repository<Color>().GetByIdAsync(request.Id);
-
-                if (colorDb == null)
-                {
-                    throw new KeyNotFoundException($"Color with ID {request.Id} not found.");
-                }
-
-                await _unitOfWork.Repository<Color>().DeleteAsync(colorDb, cancellationToken);
-
-                await _unitOfWork.SaveChangesAsync(cancellationToken);
-
-                _logger.LogInformation("Xóa màu thành công với Id: {ColorId}", request.Id);
-
-                return Unit.Value;
+                // Nếu đang được sử dụng, ném ra một ngoại lệ với thông báo rõ ràng.
+                throw new InvalidOperationException("Không thể xóa màu này vì đang có sản phẩm sử dụng.");
             }
-            catch (Exception ex)
+
+            if (colorDb == null)
             {
-                _logger.LogError(ex, "Lỗi khi xóa màu với Id: {ColorId}", request.Id);
-                throw new ApplicationException("Lỗi khi xóa màu");
+                throw new KeyNotFoundException($"Màu với id {request.Id} không tồn tại.");
             }
+
+            await _unitOfWork.Repository<Color>().DeleteAsync(colorDb, cancellationToken);
+
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Xóa màu thành công với Id: {ColorId}", request.Id);
+
+            return Unit.Value;
         }
     }
 }
